requirements.txt
anthropic==0.30.0
python-dotenv
openai==1.35.7